requirements.txt
Django>=1.4
Fabric>=1.3
South>=0.7.3
django-debug-toolbar>=0.9.4
django-devserver>=0.3.1
django-extensions>=0.7.1
django-annoying>=0.7.6
django-model-utils>=1.0.0
django-templatetag-sugar>=0.1
gunicorn>=0.13.4
